# packages/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Koji API
  version: 1.0.0-draft
  summary: Document extraction platform API — hosted cloud and self-hosted OSS.
  description: |
    The Koji API. Documents in, structured data out.

    This spec is the contract surface for the Koji product across the hosted
    platform (`api.getkoji.cloud`), the self-hosted OSS distribution, and
    enterprise / BYOC deployments. Every dashboard surface, every CLI command,
    and every public integration eventually maps to an operation defined here.

    This document is the **source of truth** for:

    - The REST API exposed by `platform/apps/hosted` (Cloudflare Workers).
    - Generated TypeScript types consumed via `@koji/types/api`.
    - Server-side request/response validation (Zod or equivalent middleware
      generated from this spec).
    - Public documentation rendered on `getkoji.dev/docs/reference/api`.

    The endpoint inventory follows `docs/specs/api-endpoints.md` §3 through §15.
    Resource shapes follow `docs/specs/database-schema.md`, stripped of fields
    that are purely internal (`tenant_id`, `deleted_at`, raw auth blobs, hashed
    secrets, etc.).

    **Status:** first-draft contract for Wave 0 of the platform roadmap. Exact
    request/response shapes are stable enough to generate clients and type
    stubs against, but may still be refined as the first endpoints land. Any
    breaking change bumps this doc and the consumers in lock-step.
  contact:
    name: Koji
...
  license:
...
servers:
...
security:
  - apiKey: []
  - sessionCookie: []
...
paths:
...
  /me:
    get:
      operationId: getCurrentUser
      tags: [Session]
      summary: Current user
      description: |
        Returns the authenticated user plus every tenant/project the user has
        access to and the roles held on each.
      responses:
        '200':
          description: Current session user.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CurrentUser'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /projects:
    get:
      operationId: listProjects
      tags: [Projects]
      summary: List projects the current user can access
      responses:
        '200':
          description: Project memberships for the current user.
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/ProjectSummary'
        '401':
          $ref: '#/components/responses/Unauthorized'
    post:
      operationId: createProject
      tags: [Projects]
      summary: Create a project
      description: Creates a new project within the current tenant. Requires `tenant-owner`.
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateProjectRequest'
      responses:
        '201':
          description: Project created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
        '429':
          $ref: '#/components/responses/RateLimited'
  /projects/{projectSlug}:
    parameters:
      - $ref: '#/components/parameters/ProjectSlug'
    get:
      operationId: getProject
      tags: [Projects]
      summary: Project detail
      responses:
        '200':
          description: Project found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
    patch:
      operationId: updateProject
      tags: [Projects]
      summary: Update project metadata
      description: Requires `project-admin`.
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateProjectRequest'
      responses:
        '200':
          description: Project updated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      operationId: deleteProject
      tags: [Projects]
      summary: Delete a project (permanent)
      description: |
        Requires `tenant-owner`. Soft-deletes the project and schedules a
        90-day window before hard deletion. **Dangerous — there is no UI
        undo beyond the recovery window.**
      responses:
        '204':
          description: Project deleted.
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /projects/{projectSlug}/schemas:
    parameters:
      - $ref: '#/components/parameters/ProjectSlug'
    get:
      operationId: listSchemas
      tags: [Schemas]
      summary: List schemas in a project
      description: |
        Powers the top-bar **schema picker** dropdown. Returns a fast, compact
        summary for each schema with current version, health, and document
        counts.
      responses:
        '200':
          description: Schemas.
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/SchemaSummary'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
    post:
      operationId: createSchema
      tags: [Schemas]
      summary: Create a schema
      description: |
        Creates a new schema within the project. Requires `schema-write`. The
        schema may be seeded from blank, from a Playground session, or from a
        public template.
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateSchemaRequest'
      responses:
        '201':
          description: Schema created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Schema'
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
  /projects/{projectSlug}/schemas/{schemaSlug}:
    parameters:
      - $ref: '#/components/parameters/ProjectSlug'
      - $ref: '#/components/parameters/SchemaSlug'
    get:
      operationId: getSchema
      tags: [Schemas]
      summary: Schema detail
      description: |
        Returns the schema plus the current version's YAML and any unsaved
        Build-mode draft.
      responses:
        '200':
          description: Schema detail.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Schema'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
    patch:
      operationId: updateSchemaDraft
      tags: [Schemas]
      summary: Update Build-mode draft YAML
      description: |
        Updates the in-progress draft YAML for this schema without committing
        a new version. Requires `schema-write`. Committing is a separate
        operation (`POST …/versions`).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateSchemaDraftRequest'
      responses:
        '200':
          description: Draft updated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Schema'
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
    delete:
      operationId: deleteSchema
      tags: [Schemas]
      summary: Archive a schema
      description: |
        Soft-delete. Pipelines that still reference the schema begin failing
        their routing stage; recover with an admin restore action within 90 days.
      responses:
        '204':
          description: Schema archived.
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
  /projects/{projectSlug}/schemas/{schemaSlug}/fields:
    parameters:
      - $ref: '#/components/parameters/ProjectSlug'
      - $ref: '#/components/parameters/SchemaSlug'
    get:
      operationId: getSchemaFields
      tags: [Schemas]
      summary: Structured field metadata for a schema
      description: |
        Returns the latest committed version's YAML normalized into a stable
        JSON shape. Lets clients (notably the review UI's override dropdowns)
        consume schema field metadata without parsing YAML on the wire.

        If the schema exists but has no committed YAML, falls back to the
        in-progress draft. If nothing is available, returns `{ fields: [] }`.
        Unknown YAML keys on a field are silently dropped (forward-compat).
      responses:
        '200':
          description: Field metadata.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SchemaFieldList'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /projects/{projectSlug}/schemas/{schemaSlug}/versions:
    parameters:
      - $ref: '#/components/parameters/ProjectSlug'
      - $ref: '#/components/parameters/SchemaSlug'
    get:
      operationId: listSchemaVersions
      tags: [Schemas]
      summary: List schema versions
      responses:
        '200':
          description: Versions.
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/SchemaVersionSummary'
        '404':
          $ref: '#/components/responses/NotFound'
    post:
      operationId: commitSchemaVersion
      tags: [Schemas]
      summary: Commit the draft as a new version
      description: |
        Promotes the current draft YAML to an immutable version (`vN+1`).
        Requires `schema-write`. The request may optionally update
        `current_version_id` to the new version.
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CommitSchemaVersionRequest'
      responses:
        '201':
          description: Version committed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SchemaVersion'
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
...
  /projects/{projectSlug}/pipelines:
    parameters:
      - $ref: '#/components/parameters/ProjectSlug'
    get:
      operationId: listPipelines
      tags: [Pipelines]
      summary: List pipelines
      description: Powers the **Pipelines** list.
      responses:
        '200':
          description: Pipelines.
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/PipelineSummary'
        '404':
          $ref: '#/components/responses/NotFound'
    post:
      operationId: createPipeline
      tags: [Pipelines]
      summary: Create a pipeline from YAML
      description: Requires `pipeline-write`.
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePipelineRequest'
      responses:
        '201':
          description: Pipeline created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Pipeline'
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
...
  /projects/{projectSlug}/jobs/{jobSlug}:
    parameters:
      - $ref: '#/components/parameters/ProjectSlug'
      - $ref: '#/components/parameters/JobSlug'
    get:
      operationId: getJob
      tags: [Jobs]
      summary: Job detail
      description: Powers the **Job detail** page.
      responses:
        '200':
          description: Job.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Job'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    apiKey:
      type: http
      scheme: bearer
      bearerFormat: kj_live_*
      description: |
        Bearer token issued via `POST /projects/{p}/api-keys`. Tokens carry a
        scope list checked in the app layer; see `docs/specs/auth-permissioning.md`.
    sessionCookie:
      type: apiKey
      in: cookie
      name: koji_session
      description: |
        Browser session cookie set by the auth adapter after `POST /auth/session`.
    webhookSignature:
      type: apiKey
      in: header
      name: X-Koji-Signature
      description: |
        HMAC-SHA256 signature over `timestamp.body` using the per-source
        secret. Format: `t=<unix>,v1=<hex>`. Used only for the inbound
        customer webhook endpoint (`/projects/{p}/sources/{s}/webhook`).
...